# src/vault/send-tips/dto/transfer.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Currency } from 'src/blockchain-interaction/currency';
import { DepositDto } from 'src/vault/deposit/dto/deposit.dto';
import { VaultDto } from 'src/vault/vault.dto';

export class TransferDto extends VaultDto {
  @ApiProperty()
  amount: number;
  @ApiProperty({ enum: Currency })
  currency: Currency;
  @ApiProperty()
  username: string;
}