# swagger.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_loginUser","summary":"Login user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/register":{"post":{"operationId":"AuthController_addUser","summary":"Register new user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/forgotPassword":{"post":{"operationId":"AuthController_forgotPassword","summary":"Forgot password","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login/admin":{"post":{"operationId":"AuthController_loginAdmin","summary":"Login Admin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/users":{"post":{"operationId":"UserController_getAllUsers","summary":"Get all users","parameters":[],"responses":{"201":{"description":""}},"tags":["Users"]}},"/users/exist":{"get":{"operationId":"UserController_checkUser","summary":"check user existring","parameters":[{"name":"username","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Users"]}},"/users/count":{"get":{"operationId":"UserController_getUsersCount","summary":"Get users count","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UserController_getUser","summary":"Get single users","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"patch":{"operationId":"UserController_updateUser","summary":"Update user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UserController_deleteUser","summary":"Delete user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/username/{username}":{"get":{"operationId":"UserController_getUserByUsername","summary":"Get single user by username","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/password/{id}":{"patch":{"operationId":"UserController_updatePassword","summary":"Update password","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/ban/{id}":{"patch":{"operationId":"UserController_banUser","summary":"Ban User","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/unban/{id}":{"patch":{"operationId":"UserController_unbanUser","summary":"Unban User","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/search":{"post":{"operationId":"UserController_searchUser","summary":"Search user","parameters":[],"responses":{"201":{"description":""}},"tags":["Users"]}},"/article":{"post":{"operationId":"ArticleController_getArticles","summary":"Get articles","parameters":[],"responses":{"201":{"description":""}},"tags":["Article"]}},"/article/count/{category}":{"get":{"operationId":"ArticleController_getArticlesCount","summary":"Get articles count","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Article"]}},"/article/add":{"post":{"operationId":"ArticleController_addArticle","summary":"Add article","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddArticleDto"}}}},"responses":{"201":{"description":""}},"tags":["Article"]}},"/article/{id}":{"delete":{"operationId":"ArticleController_deleteArticle","summary":"Delete article by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Article"]},"patch":{"operationId":"ArticleController_updateArticle","summary":"Update article by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateArticleDto"}}}},"responses":{"200":{"description":""}},"tags":["Article"]},"get":{"operationId":"ArticleController_getSingleArticleById","summary":"Get single article by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Article"]}},"/article/url/{publicUrl}":{"post":{"operationId":"ArticleController_getSingleArticleByPublicUrl","summary":"Get single article by publicUrl","parameters":[{"name":"publicUrl","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Article"]}},"/article/views/{id}":{"patch":{"operationId":"ArticleController_updateArticleViewsCount","summary":"Update article views count","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Article"]}},"/article-likes":{"post":{"operationId":"ArticleLikesController_addLike","summary":"Like article","parameters":[],"responses":{"201":{"description":""}},"tags":["ArticleLikes"]},"delete":{"operationId":"ArticleLikesController_deleteLike","summary":"Delete like","parameters":[],"responses":{"200":{"description":""}},"tags":["ArticleLikes"]}},"/category":{"get":{"operationId":"CategoryController_getCategories","summary":"Get category","parameters":[],"responses":{"200":{"description":""}},"tags":["Category"]}},"/category/add":{"post":{"operationId":"CategoryController_addCategory","summary":"Add category","parameters":[],"responses":{"201":{"description":""}},"tags":["Category"]}},"/category/{id}":{"delete":{"operationId":"CategoryController_deleteCategory","summary":"Delete category by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Category"]}},"/deposit":{"get":{"operationId":"DepositController_deposit","summary":"Make deposit","parameters":[{"name":"blockchain","required":true,"in":"query","schema":{"enum":["ethereum","binance","solana","tron","bitcoin"],"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Deposit"]}},"/balance":{"get":{"operationId":"UserBalanceController_getBalance","summary":"Get user balance for author","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserBalance"}}}}}},"tags":["User balance"]},"post":{"operationId":"UserBalanceController_getBalanceForAdmin","summary":"Get user balance for admin","parameters":[],"responses":{"201":{"description":""}},"tags":["User balance"]}},"/withdraw":{"get":{"operationId":"WithdrawController_isAvailable","summary":"is available withdraw for user right now","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Withdraw"]},"post":{"operationId":"WithdrawController_withdraw","summary":"Withdraw tokens or native crypto","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawDto"}}}},"responses":{"201":{"description":""}},"tags":["Withdraw"]}},"/withdraw/all":{"get":{"operationId":"WithdrawController_getAllWithdraws","summary":"Get all withdraws","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WithdrawDto"}}}}}},"tags":["Withdraw"]}},"/withdraw/status":{"post":{"operationId":"WithdrawController_changeStatus","summary":"Change status for withdraw","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawStatusDto"}}}},"responses":{"201":{"description":""}},"tags":["Withdraw"]}},"/send":{"post":{"operationId":"SendTipsController_sendTips","summary":"Transfer tokens or native crypto","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferDto"}}}},"responses":{"201":{"description":"Transfer successful"},"400":{"description":"Not enough balance"}},"tags":["Send to user"]}},"/articles-to-users":{"post":{"operationId":"ArticlesToUsersController_addArticleToUser","summary":"Add article to user","parameters":[],"responses":{"201":{"description":""}},"tags":["Article To Users"]}}},"info":{"title":"Binobi","description":"The Binobi API description","version":"1.0","contact":{}},"tags":[{"name":"Binobi","description":""}],"servers":[],"components":{"schemas":{"LoginDto":{"type":"object","properties":{"login":{"type":"string","description":"login"},"password":{"type":"string","description":"Password"}},"required":["login","password"]},"RegisterDto":{"type":"object","properties":{"username":{"type":"string","description":"username"},"email":{"type":"string","description":"email"},"password":{"type":"string","description":"Password"}},"required":["username","email","password"]},"ForgotPasswordDto":{"type":"object","properties":{"userId":{"type":"string","description":"user id"},"userData":{"type":"string","description":"user data"}},"required":["userId","userData"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string","description":"username"},"email":{"type":"string","description":"email"},"password":{"type":"string","description":"Password"},"appleId":{"type":"string","description":"appleId"},"isVerified":{"type":"boolean","description":"isVerified"}}},"AddArticleDto":{"type":"object","properties":{"title":{"type":"string","description":"title"},"title_ru":{"type":"string","description":"title_ru"},"title_ua":{"type":"string","description":"title_ua"},"title_de":{"type":"string","description":"title_de"},"title_es":{"type":"string","description":"title_es"},"title_fr":{"type":"string","description":"title_fr"},"title_pt":{"type":"string","description":"title_pt"},"title_tr":{"type":"string","description":"title_tr"},"text":{"type":"text","description":"text"},"text_ru":{"type":"text","description":"text_ru"},"text_ua":{"type":"text","description":"text_ua"},"text_de":{"type":"text","description":"text_de"},"text_es":{"type":"text","description":"text_es"},"text_fr":{"type":"text","description":"text_fr"},"text_pt":{"type":"text","description":"text_pt"},"text_tr":{"type":"text","description":"text_tr"},"subtitle":{"type":"text","description":"text"},"subtitle_ru":{"type":"text","description":"text_ru"},"subtitle_ua":{"type":"text","description":"text_ua"},"subtitle_de":{"type":"text","description":"subtitle_de"},"subtitle_es":{"type":"text","description":"subtitle_es"},"subtitle_fr":{"type":"text","description":"subtitle_fr"},"subtitle_pt":{"type":"text","description":"subtitle_pt"},"subtitle_tr":{"type":"text","description":"subtitle_tr"},"publicUrl":{"type":"string","description":"public url"},"categoryId":{"type":"string","description":"array of ids"}},"required":["title","title_ru","title_ua","title_de","title_es","title_fr","title_pt","title_tr","text","text_ru","text_ua","text_de","text_es","text_fr","text_pt","text_tr","subtitle","subtitle_ru","subtitle_ua","subtitle_de","subtitle_es","subtitle_fr","subtitle_pt","subtitle_tr","publicUrl","categoryId"]},"UpdateArticleDto":{"type":"object","properties":{"title":{"type":"string","description":"title"},"title_ru":{"type":"string","description":"title_ru"},"title_ua":{"type":"string","description":"title_ua"},"title_de":{"type":"string","description":"title_de"},"title_es":{"type":"string","description":"title_es"},"title_fr":{"type":"string","description":"title_fr"},"title_pt":{"type":"string","description":"title_pt"},"title_tr":{"type":"string","description":"title_tr"},"text":{"type":"text","description":"text"},"text_ru":{"type":"text","description":"text_ru"},"text_ua":{"type":"text","description":"text_ua"},"text_de":{"type":"text","description":"text_de"},"text_es":{"type":"text","description":"text_es"},"text_fr":{"type":"text","description":"text_fr"},"text_pt":{"type":"text","description":"text_pt"},"text_tr":{"type":"text","description":"text_tr"},"subtitle":{"type":"text","description":"subtitle"},"subtitle_ru":{"type":"text","description":"subtitle_ru"},"subtitle_ua":{"type":"text","description":"subtitle_ua"},"subtitle_de":{"type":"text","description":"subtitle_de"},"subtitle_es":{"type":"text","description":"subtitle_es"},"subtitle_fr":{"type":"text","description":"subtitle_fr"},"subtitle_pt":{"type":"text","description":"subtitle_pt"},"subtitle_tr":{"type":"text","description":"subtitle_tr"},"publicUrl":{"type":"string","description":"public url"},"categoryId":{"type":"string","description":"array of ids"}},"required":["title","title_ru","title_ua","title_de","title_es","title_fr","title_pt","title_tr","text","text_ru","text_ua","text_de","text_es","text_fr","text_pt","text_tr","subtitle","subtitle_ru","subtitle_ua","subtitle_de","subtitle_es","subtitle_fr","subtitle_pt","subtitle_tr","publicUrl","categoryId"]},"UserBalance":{"type":"object","properties":{"blockchain":{"type":"string","enum":["ethereum","binance","solana","tron","bitcoin"]},"currency":{"type":"string","enum":["usdt","native"],"description":"native or another token"},"amount":{"type":"number"}},"required":["blockchain","currency","amount"]},"WithdrawDto":{"type":"object","properties":{"blockchain":{"type":"string","enum":["ethereum","binance","solana","tron","bitcoin"]},"amount":{"type":"number"},"currency":{"type":"string","enum":["usdt","native"]},"address":{"type":"string"}},"required":["blockchain","amount","currency","address"]},"WithdrawStatusDto":{"type":"object","properties":{"status":{"type":"string","enum":["PENDING","FULFILLED","REJECTED"]},"id":{"type":"string"}},"required":["status","id"]},"TransferDto":{"type":"object","properties":{"blockchain":{"type":"string","enum":["ethereum","binance","solana","tron","bitcoin"]},"amount":{"type":"number"},"currency":{"type":"string","enum":["usdt","native"]},"username":{"type":"string"}},"required":["blockchain","amount","currency","username"]}}}}